+++ AllMyLights/Transformations/ColorTransformation.cs
@@ -3,11 +3,14 @@ using System.Drawing;
 using System.Reactive.Linq;
 using AllMyLights.Common;
 using AllMyLights.Models.Transformations;
+using NLog;
 
 namespace AllMyLights.Transformations
 {
     public class ColorTransformation: ITransformation<Ref<Color>>
     {
+        private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
+
         public const string Type = "Color";
 
         private string ChannelLayout { get; }
@@ -27,7 +30,11 @@ namespace AllMyLights.Transformations
                         throw new ArgumentException($"{nameof(ColorTransformation)} requires input to be of type string");
                     }
 
-                    return new Ref<Color>(ColorConverter.Decode(input as string, ChannelLayout));
+                    Logger.Debug($"Decoding color from {input}");
+                    Ref<Color> colorRef = new Ref<Color>(ColorConverter.Decode(input as string, ChannelLayout));
+                    Logger.Debug($"Decoded {colorRef.Value}");
+
+                    return colorRef;
                 });
             };
         }