+++ AllMyLights/Transformations/TransformationFactory.cs
@@ -10,6 +10,7 @@ namespace AllMyLights.Transformations
             return options switch
             {
                 ColorTransformationOptions colorOptions => new ColorTransformation(colorOptions),
+                MappingTransformationOptions mappingOptions => new MappingTransformation(mappingOptions),
                 JsonPathTransformationOptions jpathOptions => new JsonPathTransformation<string>(jpathOptions),
                 _ => throw new NotImplementedException($"Transformation for type {options.Type} not registered")
             };